move builder to be publicly accessible
